\name{birth_2006}
\alias{birth_2006}
\docType{data}
\title{
  Half of male births recorded in the Comunitat Valenciana (Spain) during 2006.
}
\description{
  Data frame containing the dates of birth of a random sample composed by half of all male babies born in the region of Comunitat Valenciana (Spain) during 2006.
}
\usage{data(birth_2006)}

\format{A data frame containing 13,626 rows and 1 column:
    \describe{
      \item{date.birth}{Dates of birth in format "yyyy-mm-dd".}
    }}

\source{
   National Statistical Institute, \url{https://www.ine.es/}.
}

\keyword{datasets}
