\name{death_2006}
\alias{death_2006}
\docType{data}
\title{
  Half of male deaths recorded in Comunitat Valenciana (Spain) during year 2006.
}
\description{
  Data frame containing the dates of birth and death of a random sample composed by half of all men who died in the region of Comunitat Valenciana (Spain) during the year 2006.
}
\usage{data(death_2006)}

\format{A data frame containing 10,146 rows and 2 columns:
    \describe{
      \item{date.birth}{Dates of birth in format "yyyy-mm-dd".}
      \item{date.death}{Dates of death in format "yyyy-mm-dd".}
    }}

\source{
    National Statistical Institute, \url{https://www.ine.es/}.
}

\keyword{datasets}
