\name{emi_2006}
\alias{emi_2006}
\docType{data}
\title{
  Male emigrants recorded in Comunitat Valenciana (Spain) during year 2006.
}
\description{
  Data frame containing the dates of birth and emigration of a random sample composed by half of all men who emigrated from the region of Comunitat Valenciana (Spain) during the year 2006.
}
\usage{data(emi_2006)}

\format{A data frame containing 18,006 rows and 2 columns:
    \describe{
      \item{date.birth}{Dates of birth in format "yyyy-mm-dd".}
      \item{date.emi}{Dates of death in format "yyyy-mm-dd".}
    }}

\source{
    National Statistical Institute, \url{https://www.ine.es/}.
}

\keyword{datasets}
