\name{immi_2006}
\alias{immi_2006}
\docType{data}
\title{
  Half of the male immigrants recorded in Comunitat Valenciana (Spain) during year 2006.
}
\description{
  Data frame containing the dates of birth and immigration of a random sample composed by half of all men who immigrate to the region of Comunitat Valenciana (Spain) during the year 2006.
}
\usage{data(immi_2006)}

\format{A data frame containing 48,431 rows and 2 columns:
    \describe{
      \item{date.birth}{Dates of birth in format "yyyy-mm-dd".}
      \item{date.immi}{Dates of death in format "yyyy-mm-dd".}
    }}

\source{
    National Statistical Institute, \url{https://www.ine.es/}.
}

\keyword{datasets}
