\name{pop_2006}
\alias{pop_2006}
\docType{data}
\title{
  Half of the male population living in the Comunitat Valenciana (Spain) on January, 1 2006.
}
\description{
  Data frame containing the dates of birth of a random sample composed by half of all men who were alive on 1 January 2006 in the region of Comunitat Valenciana (Spain).
}
\usage{data(pop_2006)}

\format{A data frame containing 1,197,154 rows and 1 column:
    \describe{
      \item{date.birth}{Dates of birth in format "yyyy-mm-dd".}
    }}

\source{
  Instituto Valenciano de Estadistica (IVE) \url{https://pegv.gva.es/es/}.
}

\keyword{datasets}
